+++ test/mocks/http/mocks.h
@@ -207,7 +207,8 @@ public:
 
   // Http::StreamDecoderFilterCallbacks
   void sendLocalReply(Code code, const std::string& body,
-                      std::function<void(HeaderMap& headers)> modify_headers) override {
+                      std::function<void(HeaderMap& headers)> modify_headers,
+                      bool rate_limited_as_resource_exhausted) override {
     Utility::sendLocalReply(
         is_grpc_request_,
         [this, modify_headers](HeaderMapPtr&& headers, bool end_stream) -> void {
@@ -217,7 +218,7 @@ public:
           encodeHeaders(std::move(headers), end_stream);
         },
         [this](Buffer::Instance& data, bool end_stream) -> void { encodeData(data, end_stream); },
-        stream_destroyed_, code, body, is_head_request_);
+        stream_destroyed_, code, body, is_head_request_, rate_limited_as_resource_exhausted);
   }
   void encode100ContinueHeaders(HeaderMapPtr&& headers) override {
     encode100ContinueHeaders_(*headers);